move the analyzer to a subpackage
